Commentaire et message d'erreurs

+++ src/AppBundle/Controller/ArmyController.php
@@ -10,6 +10,7 @@ use AppBundle\Form\Army\ArmyType;
 
 class ArmyController extends Controller
 {
+    // Accueil de la zone armée
     public function indexAction()
     {
         $em = $this->getDoctrine()->getManager();
@@ -19,25 +20,23 @@ class ArmyController extends Controller
 
         return $this->render('AppBundle:Army:index.html.twig', array('listArmies' => $listArmies, 'listUsers' => $listUsers));
     }
-
-    public function viewAction($slugArmy)
+    // Page de vue d'une armée
+    public function viewAction(Army $army)
     {
-        $em = $this->getDoctrine()->getManager();
-
-        $army = $em->getRepository('AppBundle:Army\Army')->FindOneWithAll($slugArmy);
-
+        // On vérifie si $army existe
         if ($army === null) {
             throw new NotFoundHttpException('Armée inexistante');
         }
-
+        // On récupere la liste des groupes
         $listGroupes = $this->getDoctrine()
-            ->getManager()
-            ->getRepository('AppBundle:Army\Groupe')->findAll();
+                            ->getManager()
+                            ->getRepository('AppBundle:Army\Groupe')->findAll();
 
         return $this->render('AppBundle:Army:view.html.twig', array('army' => $army,
             'listgroupes' => $listGroupes, ));
     }
 
+    //Page de création d'une armée
     public function addAction(Request $request)
     {
         $em = $this->getDoctrine()->getManager();
@@ -51,25 +50,36 @@ class ArmyController extends Controller
             $em->persist($army);
             $em->flush();
 
-            $request->getSession()->getFlashBag()->add('info', 'Votre armée a été créé avec succès.');
-
-            return $this->redirectToRoute('army_view', array('slugArmy' => $army->getSlugArmy()));
+            return $this->redirectToRoute('armies');
         }
 
         return $this->render('AppBundle:Army:add.html.twig', array('form' => $form->createView()));
     }
+
+    // Page de modifiaction de l'armée
     public function editAction(Request $request, Army $army)
     {
+        // Vérification si $army existe
+        if(null === $army)
+        {
+            throw new NotFoundHttpException('Armée inexistant.');
+        }
+
+        // On s'assure que l'utilisateur est le propriétaire de l'armée
+        if( $army->getUser() !== $this->get('security.token_storage')->getToken()->getUser()){
+            $request->getSession()->getFlashBag()->add('danger', 'Vous ne pouvez pas supprimer cette armée !');
+            return $this->redirectToRoute('armies');
+        }
         $em = $this->getDoctrine()->getManager();
 
+        // On utilise le même formulaire que pour la création d'une armée mais on enleve le champ race qui ne peut être modifier
         $form = $this->createForm(ArmyType::class, $army);
 
         $form->remove('race');
 
         if ($request->isMethod('POST') && $form->handleRequest($request)->isValid()) {
-
             $em->flush();
-            $request->getSession()->getFlashbag()->add('info', 'Votre armée a été modifié avec succès.');
+            $request->getSession()->getFlashbag()->add('info', 'Votre armée a bien été modifier');
 
             return $this->redirectToRoute('army_view', array('slugArmy' => $army->getSlugArmy()));
         }
@@ -77,17 +87,28 @@ class ArmyController extends Controller
         return $this->render('AppBundle:Army:edit.html.twig', array('form' => $form->createView(), 'army' => $army));
     }
 
+    // Page de suppresion d'armée
     public function deleteAction(Request $request, Army $army)
     {
+        // Vérification si $army existe
+        if(null === $army)
+        {
+            throw new NotFoundHttpException('Armée inexistant');
+        }
+        // On s'assure que l'utilisateur est le propriétaire de l'armée
+        if( $army->getUser() !== $this->get('security.token_storage')->getToken()->getUser()){
+            $request->getSession()->getFlashBag()->add('danger', 'Vous ne pouvez pas supprimer cette armée !');
+            return $this->redirectToRoute('armies');
+        }
+        // On crée un form pour supprimer l'armée
         $em = $this->getDoctrine()->getManager();
-
         $form = $this->get('form.factory')->create();
 
         if ($request->isMethod('POST') && $form->handleRequest($request)->isValid()) {
             $em->remove($army);
             $em->flush();
 
-            $request->getSession()->getFlashBag()->add('info', 'Votre armée a été supprimé avec succès.');
+            $request->getSession()->getFlashBag()->add('info', 'Votre armée a bien été supprimé !');
 
             return $this->redirectToRoute('armies');
         }
